v1.0.0: Migración completa a enum Rol y resolución de errores
- ✅ Actualizado CHANGELOG a versión 1.0.0
- ✅ Migrado sistema de roles de String a enum Rol con métodos utilitarios
- ✅ Eliminado UserRole class, reemplazado por enum Rol
- ✅ Actualizado DTOs con soporte completo para enum y helper methods
- ✅ Refactorizado servicios para usar enum en lugar de constantes
- ✅ Corregido DataInitializer para resolver ClassCastException
- ✅ Solucionado GlobalExceptionHandler para eliminar errores 500
- ✅ Configurado SpringDoc OpenAPI correctamente (/v3/api-docs)
- ✅ Mejorado manejo de excepciones y documentación Swagger
- ✅ Agregada documentación completa del proyecto y configuración

+++ src/main/java/com/poo/miapi/dto/auth/ChangePasswordDto.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 public class ChangePasswordDto {
 
     @NotNull(message = "El ID del usuario no puede ser nulo")
-    private Long userId;
+    private int userId;
 
     @NotBlank(message = "La nueva contraseña no puede estar vacía")
     private String newPassword;
@@ -14,16 +14,16 @@ public class ChangePasswordDto {
     public ChangePasswordDto() {
     }
 
-    public ChangePasswordDto(Long userId, String newPassword) {
+    public ChangePasswordDto(int userId, String newPassword) {
         this.userId = userId;
         this.newPassword = newPassword;
     }
 
-    public Long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
